LCOM1 fix

+++ architecture-evaluator-backend/src/main/java/io/github/Opsord/architecture_evaluator_backend/modules/parser/detailer/dto/parts/CohesionMetricsDTO.java
@@ -5,7 +5,7 @@ import lombok.Data;
 @Data
 public class CohesionMetricsDTO {
     private Integer couplingBetweenObjects; // CBO: number of classes that this class depends on
-    private Double lackOfCohesion1;
-    private Double lackOfCohesion2;
+    private Integer lackOfCohesion1;
+    private Integer lackOfCohesion2;
     private Double cohesionAmongMethods;
 }
